requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-dotenv==1.0.0
qdrant-client==1.7.0
pandas==2.2.2
fasttext==0.9.3
openpyxl==3.1.5
beautifulsoup4==4.12.3
tqdm==4.66.5
numpy>=1.22.4,<2.0.0
langchain==0.1.12
openai==1.14.2

mlflow==2.11.1
pyyaml
nltk==3.9.1
numpy
scipy
transformers==4.44.2
whoosh==2.7.4
contractions==0.1.73
langdetect==1.0.9
